fix!: improve device management
The behavior of --simulator and --emulator options changed.
They are no longer mandatory. When not passed, the CLI attempts
to find the booted device. A device that is already booted is not
shutdown by the CLI before exit. If none or more than one is found,
the CLI exists with an error. If either option is passed, unless
the device is not installed in the system, the CLI launches the test
app into the selected device. In this case, the device is booted and
shutdown by the CLI automatically.

It's now possible to have multiple AVDs running simultaneously and
use any of them to install the test app into.

CLI options now override those defined in the configuration file.